+++ changelog.md
@@ -1,5 +1,6 @@
 # Change Log of SilverShop JSON Response
 
 * 3.0.0 update for SilverShop v3
+  * Added Modifiers to response data
 * 2.0.0 Update for SilverShop v2
 * 1.0.0 Initial Release